[Scalings] Restructuring scalings functions
