+++ objectPerformance/src/plotter.py
@@ -6,10 +6,10 @@ import matplotlib.pyplot as plt
 import mplhep as hep
 import numpy as np
 from progress.bar import IncrementalBar
-from scipy.optimize import curve_fit
 import yaml
 
 from turnon_collection import TurnOnCollection
+from scaling_collection import ScalingCollection
 import utils
 
 
@@ -18,10 +18,10 @@ plt.style.use(hep.style.CMS)
 
 class Plotter():
 
-    def _make_output_dirs(self):
-        os.makedirs("outputs/turnons", exist_ok=True)
-        os.makedirs("outputs/distributions", exist_ok=True)
-        os.makedirs("outputs/scalings", exist_ok=True)
+    def _make_output_dirs(self, version: str):
+        os.makedirs(f"outputs/{version}/turnons", exist_ok=True)
+        os.makedirs(f"outputs/{version}/distributions", exist_ok=True)
+        os.makedirs(f"outputs/{version}/scalings", exist_ok=True)
 
     def _create_new_plot(self):
         fig, ax = plt.subplots(figsize=(10, 10))
@@ -35,6 +35,7 @@ class EfficiencyPlotter(Plotter):
         self.plot_name = name
         self.cfg = cfg
         self.turnon_collection = turnon_collection
+        self.version = self.turnon_collection.version
         self.bin_width = turnon_collection.cfg_plot.bin_width
 
     def _style_plot(self, fig, ax, legend_loc="lower right"):
@@ -74,7 +75,7 @@ class EfficiencyPlotter(Plotter):
 
         self._style_plot(fig, ax)
         ax.set_ylim(0, 1.1)
-        plt.savefig(f"outputs/turnons/{self.plot_name}_"
+        plt.savefig(f"outputs/{self.version}/turnons/{self.plot_name}_"
                     f"{self.turnon_collection.threshold}.png")
         plt.close()
 
@@ -105,12 +106,12 @@ class EfficiencyPlotter(Plotter):
                         color=test_hist[0].get_color(), **err_kwargs)
 
         self._style_plot(fig, ax)
-        plt.savefig(f"outputs/distributions/{self.plot_name}"
+        plt.savefig(f"outputs/{self.version}/distributions/{self.plot_name}"
                     f"_{self.turnon_collection.threshold}_dist.png")
         plt.close()
 
     def plot(self):
-        self._make_output_dirs()
+        self._make_output_dirs(self.version)
         self._plot_efficiency_curve()
         self._plot_raw_counts()
 
@@ -159,11 +160,12 @@ class EfficiencyCentral():
 class ScalingPlotter(Plotter):
 
     def __init__(self, plot_name: str, cfg_plot: dict, scalings: dict,
-                 scaling_pct: float, params: dict):
+                 scaling_pct: float, version: str, params: dict):
         self.plot_name = plot_name
         self.cfg_plot = cfg_plot
         self.scalings = scalings
         self.params = params
+        self.version = version
         self.scaling_pct = scaling_pct
 
     def _params_to_func_str(self, obj):
@@ -184,7 +186,7 @@ class ScalingPlotter(Plotter):
         ax.set_ylim(0, ymax)
 
     def plot(self):
-        self._make_output_dirs()
+        self._make_output_dirs(self.version)
 
         fig, ax = self._create_new_plot()
         for obj, points in self.scalings.items():
@@ -206,7 +208,7 @@ class ScalingPlotter(Plotter):
         self._set_plot_ranges(ax)
         fig.tight_layout()
 
-        plt.savefig(f"outputs/scalings/{self.plot_name}.png")
+        plt.savefig(f"outputs/{self.version}/scalings/{self.plot_name}.png")
         plt.close()
 
 
@@ -233,262 +235,20 @@ class ScalingCentral():
             "Failed to find thresholds in cfg_scaling_thresholds!"
         )
 
-    def _find_percentage_point(self, hist, bins, scaling_pct):
-        for i, eff in enumerate(hist[:-2]):
-            is_point = (
-                eff < scaling_pct
-                and hist[i + 1] > scaling_pct
-                and hist[i + 2] > scaling_pct
-            )
-            if is_point:
-                return bins[i + 1]
-
-    def _find_turnon_cut(self, graph_x, graph_y, Target):
-        L = 0
-        R = np.max(graph_x)
-
-        while (R - L > 0.0001):
-            C = (L + R) / 2
-            V = self._get_point_on_curve(C, graph_x, graph_y)
-
-            if (V < Target):
-                L = C
-            else:
-                R = C
-
-        return (R + L) / 2.
-
-    def _find_turnon_fit(self, function, popt, _min, _max, target):
-        if ((function(_min, *popt) > target)
-           | (function(_max, *popt) < target)):
-            return -1000
-
-        L = _min
-        R = _max
-
-        while (R - L > 0.0001):
-            C = (L + R) / 2
-            V = function(C, *popt)
-
-            if (V < target):
-                L = C
-            else:
-                R = C
-
-        return (R + L) / 2
-
-    @utils.ignore_warnings
-    def _compute_value_of_tanh_at_threshold(self,
-                                            efficiency,
-                                            bins,
-                                            threshold):
-        xvals = np.array(bins)
-        efficiency = np.array(efficiency)
-
-        xvals = xvals[~np.isnan(efficiency)]
-        efficiency = efficiency[~np.isnan(efficiency)]
-
-        popt, pcov = curve_fit(utils.tanh, xvals, efficiency, p0=[1, 0])
-        if np.inf in pcov:
-            return None
-
-        s_val = utils.arctanh(threshold, *list(popt))
-
-        return s_val
-
-    @utils.ignore_warnings
-    def _compute_value_of_errf_at_threshold(self,
-                                            efficiency,
-                                            bins,
-                                            scaling_pct):
-        xvals = np.array(bins)
-        efficiency = np.array(efficiency)
-
-        xvals = xvals[~np.isnan(efficiency)]
-        efficiency = efficiency[~np.isnan(efficiency)]
-
-        popt, pcov = curve_fit(utils.errf, xvals, efficiency,
-                               p0=[0.02, 80, 20, 1, np.min(efficiency)]
-                               )
-
-        s_val = self._find_turnon_fit(utils.errf, popt,
-                                      np.min(xvals), np.max(xvals) * 10,
-                                      scaling_pct * utils.errf(10000, *popt)
-                                      )
-
-        return s_val
-
-    def _interpolate(self, H, K1, K2):
-
-        A = np.ones(len(K1)) * (-K2)
-        B = [k1i + 2 * K2 for k1i in K1]
-        C = np.ones(len(K1)) * (-K2)
-        D = []
-
-        A[0] = 0
-        C[-1] = 0
-
-        D = [k1i * h1i for k1i, h1i in zip(K1, H)]
-
-        D[0] = D[0] + K2 * 0
-        D[-1] = D[-1] + K2 * 1
-
-        for i in range(1, len(K1)):
-            F = A[i] / B[i - 1]
-
-            A[i] = A[i] - B[i - 1] * F
-            B[i] = B[i] - C[i - 1] * F
-            C[i] = C[i]
-            D[i] = D[i] - D[i - 1] * F
-
-        Y = np.ones(len(K1))
-        Y[-1] = D[-1] / B[-1]
-
-        for i in reversed(range(len(K1) - 2)):
-            Y[i] = (D[i] - C[i] * Y[i + 1]) / B[i]
-
-        return Y
-
-    def _get_point_on_curve(self, x, graph_x, graph_y):
-
-        if (x < graph_x[0]):
-            return 0
-
-        if (x >= graph_x[len(graph_x) - 1]):
-            return 1
-
-        xr = graph_x[0]
-        yr = graph_y[0]
-        for i in range(len(graph_x) - 1):
-            xl = xr
-            yl = yr
-            xr = graph_x[i + 1]
-            yr = graph_y[i + 1]
-            if ((x < xr) & (x >= xl)):
-                return yl + (yr - yl) / (xr - xl) * (x - xl)
-
-        return -1
-
-    def _compute_scalings_naive(self, turnon_collection, scalings,
-                                scaling_pct):
-        bins = turnon_collection.bins
-        bins = 0.5 * (bins[1:] + bins[:-1])
-        threshold = turnon_collection.threshold
-
-        for obj_key, gen_hist_trig in turnon_collection.hists.items():
-            if obj_key == "ref":
-                continue
-            efficiency, yerr = turnon_collection.get_efficiency(obj_key)
-
-            xbins = bins
-            xbins = xbins[~np.isnan(efficiency)]
-            er_dn = yerr[0]
-            er_up = yerr[1]
-            er_dn = er_dn[~np.isnan(efficiency)]
-            er_up = er_up[~np.isnan(efficiency)]
-            efficiency = efficiency[~np.isnan(efficiency)]
-
-            K1 = []
-            for i in range(len(efficiency)):
-                K1.append(1 / (er_dn[i] + er_up[i]) / (er_up[i] + er_dn[i]))
-
-            percentage_point = self._find_turnon_cut(
-                xbins,
-                self._interpolate(efficiency, K1, 100),
-                scaling_pct
-            )
-            if percentage_point:
-                scalings[obj_key][threshold] = percentage_point
-
-        return scalings
-
-    def _compute_scalings_tanh(self,
-                               turnon_collection,
-                               scalings,
-                               scaling_pct):
-        bins = turnon_collection.bins
-        bins = 0.5 * (bins[1:] + bins[:-1])
-        threshold = turnon_collection.threshold
-
-        for obj_key, gen_hist_trig in turnon_collection.hists.items():
-            if obj_key == "ref":
-                continue
-            efficiency, _ = turnon_collection.get_efficiency(obj_key)
-            percentage_point = self._compute_value_of_tanh_at_threshold(
-                efficiency,
-                bins,
-                scaling_pct
-            )
-            if percentage_point:
-                scalings[obj_key][threshold] = percentage_point
-
-        return scalings
-
-    def _compute_scalings_errf(self,
-                               turnon_collection,
-                               scalings,
-                               scaling_pct):
-        bins = turnon_collection.bins
-        bins = 0.5 * (bins[1:] + bins[:-1])
-        threshold = turnon_collection.threshold
-
-        for obj_key, gen_hist_trig in turnon_collection.hists.items():
-            if obj_key == "ref":
-                continue
-            efficiency, _ = turnon_collection.get_efficiency(obj_key)
-            percentage_point = self._compute_value_of_errf_at_threshold(
-                efficiency,
-                bins,
-                scaling_pct
-            )
-            if percentage_point:
-                scalings[obj_key][threshold] = percentage_point
-
-        return scalings
-
-    def _compute_scalings(self, turnon_collection, scalings,
-                          scaling_pct, method="tanh") -> dict:
-        if method == "tanh":
-            return self._compute_scalings_tanh(
-                turnon_collection,
-                scalings,
-                scaling_pct
-            )
-        if method == "errf":
-            return self._compute_scalings_errf(
-                turnon_collection,
-                scalings,
-                scaling_pct
-            )
-        if method == "naive":
-            return self._compute_scalings_naive(
-                turnon_collection,
-                scalings,
-                scaling_pct
-            )
-
-    def _fit_linear_functions(self, scalings):
-        params = {}
-        for obj, thresh_points in scalings.items():
-            xdata = [th for th, val in thresh_points.items() if val]
-            ydata = [thresh_points[x] for x in xdata]
-            if not ydata:
-                return None
-            popt, pcov = curve_fit(utils.scaling_func, xdata, ydata)
-            params[obj] = popt
-        return params
-
     def _rate_config_function(self, name: str, a: float, b: float):
         pm = '+' if b > 0 else ''
         f_string = (f"function :: {name}Scaling :: "
                     f"args:=(offline); lambda:={a}*offline{pm}{b}")
         return f_string
 
-    def _write_scalings_to_file(self, plot_name: str, params: dict):
-        with open(f"outputs/scalings/{plot_name}.txt", 'w+') as f:
+    def _write_scalings_to_file(self,
+                                plot_name: str,
+                                version: str,
+                                params: dict):
+        with open(f"outputs/{version}/scalings/{plot_name}.txt", 'w+') as f:
             f.write('')
 
-        with open(f"outputs/scalings/{plot_name}.txt", 'a') as f:
+        with open(f"outputs/{version}/scalings/{plot_name}.txt", 'a') as f:
             for obj, obj_params in params.items():
                 a, b = obj_params
                 f.write(self._rate_config_function(obj, a, b) + "\n")
@@ -509,19 +269,23 @@ class ScalingCentral():
                 turnon_collection.create_hists()
                 scaling_pct = turnon_collection.cfg_plot.scaling_pct
                 method = turnon_collection.cfg_plot.scaling_method
-                scalings = self._compute_scalings(turnon_collection,
-                                                  scalings,
-                                                  scaling_pct,
-                                                  method)
+                scaling_collect = ScalingCollection(cfg_plot,
+                                                    method,
+                                                    scaling_pct)
+                version = turnon_collection.version
+                scalings = scaling_collect._compute_scalings(turnon_collection,
+                                                             scalings,
+                                                             scaling_pct,
+                                                             method)
 
             bar.finish()
 
-            params = self._fit_linear_functions(scalings)
+            params = scaling_collect._fit_linear_functions(scalings)
             if params:
                 plotter = ScalingPlotter(plot_name, cfg_plot, scalings,
-                                         scaling_pct, params)
+                                         scaling_pct, version, params)
                 plotter.plot()
-                self._write_scalings_to_file(plot_name, params)
+                self._write_scalings_to_file(plot_name, version, params)
 
 
 if __name__ == "__main__":
